feat(c-api) Move `table` into its own module.
