+++ lib/c-api/src/wasm_c_api/externals/mod.rs
@@ -1,2 +1,3 @@
 pub mod global;
 pub mod memory;
+pub mod table;
